% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalSamplingPulp.R
\name{conditionalSamplingPulp}
\alias{conditionalSamplingPulp}
\title{Conditional sampling}
\usage{
conditionalSamplingPulp(saveDirName, pulp_cond, pulp_raw)
}
\arguments{
\item{saveDirName}{Path to the place where previous characterization was saved}

\item{conditioningData}{New data which does not contain the wall thickness or external fibrillation}
}
\description{
conditionalSamplingPulp() takes as input a dataframe
}
\examples{
conditionalSamplingPulp()

}
