% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMarginFunction.R
\name{fitMarginFunction}
\alias{fitMarginFunction}
\title{A pulpDistr function}
\usage{
fitMarginFunction(observedMargin, listOfModels, quantilesToEval)
}
\arguments{
\item{observedMargin}{ECDF values}

\item{listOfModels}{Univariate margin families to evaluate}

\item{quantilesToEval}{Quantiles to calculate values for}
}
\description{
A pulpDistr function
}
\examples{
fitMarginFunction()

}
