% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePseudoU.R
\name{pseudoObs}
\alias{pseudoObs}
\title{A pulpDistr function}
\usage{
pseudoObs(observedSample)
}
\arguments{
\item{observedSample}{Matrix or dataframe of observations}
}
\description{
pseudoObs() takes as input a matrix or dataframe of observations
and returns a matrix of ranked pseudo-observations
[0, 1)^d where d is the number of columns in the input.
}
\details{
The observations are normalized such that the value 1 is never
returned, as this could generate some numerical difficulties
later on.
}
\examples{
fitMarginFunction()

}
