% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePseudoU.R
\name{pseudoObs}
\alias{pseudoObs}
\title{A pulpDistr function}
\usage{
pseudoObs(observedSample)
}
\arguments{
\item{observedSample}{Matrix or dataframe of observations}
}
\description{
A pulpDistr function
}
\examples{
fitMarginFunction()

}
